+++ pipe_test.go
@@ -7,7 +7,7 @@ import (
 	"testing"
 
 	"pipelined.dev/pipe"
-	"pipelined.dev/pipe/internal/mock"
+	"pipelined.dev/pipe/mock"
 	"pipelined.dev/pipe/mutability"
 )
